fix: explicitly state that it is int64 (#101)

+++ future_inverse_perpetual.go
@@ -72,7 +72,7 @@ type PremiumIndexKlineResult struct {
 type PremiumIndexKlineParam struct {
 	Symbol   SymbolFuture `url:"symbol"`
 	Interval Interval     `url:"interval"`
-	From     int          `url:"from"`
+	From     int64        `url:"from"`
 
 	Limit *int `url:"limit,omitempty"`
 }
